Pre-Significant Change Commit
I realised that with my current method, there would be some issues with trying to connect few points to many points. A solution is to increase the resolution in parts of the mesh to be able to link up to the extra points. The problem with this is that the resolution increase will propagate down the branches until there are a ludicrous amount of points/faces. To reduce the impact of this, I intend to only add points to the circle when needed.