Implemented turtle for creating trees out of lines
I have created a turtle with a few commands and a language to give the turtle commands.
I am going to have to backtrack with the circle. I have worked out that the first way I had did work and this version does not. Unfortunately, I did not separate it the turtle onto a different branch so I will have to start again