Finish minimum viable product
I made a lot of changes toward finishing this project without committing to git. What I think let this project down was not using git as much as I could have. I didn't use the branch system for new features and that really would have helped when I decided I wanted to go back on one feature but not another